Create noise.c
This corrupts a .wav file by XOR-ing each floating point value of the audio with the key passed as an argument to the script. The results are often just harsh unpredictable noise since bitwise operations are done on floating point values, which can result in extreme values. This serves no real functional purpose since the product audio file is corrupt, but it cured my curiosity :)